Intersection method.

+++ src/drone/Drone.java
@@ -10,7 +10,7 @@ public interface Drone {
 
     void hover();
 
-    void move(int x, int y, int z);
+    void move(float x, float y, float z);
 
     void changeLightVolume(int percentage);
