+++ src/formation/Coordinate.java
@@ -1,11 +1,11 @@
 package formation;
 
 public class Coordinate {
-    public int x;
-    public int y;
-    public int z;
+    public float x;
+    public float y;
+    public float z;
 
-    public Coordinate(int x, int y, int z){
+    public Coordinate(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
